pyproject.toml
[project]
name = "tiny-lightning-hydra-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.12"
dependencies = [
    "hydra-core==1.3.2",
    "lightning==2.5.5",
    "pandas==2.3.2",
    "rootutils==1.0.7",
    "torch==2.8.0",
    "torchmetrics==1.8.2",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "wandb==0.22.0",
    "ipykernel==6.30.1",
    "timm==1.0.20",
    "diffusers==0.35.1",
    "numpy==2.0.2",
    "matplotlib==3.9.4",
    "scipy==1.13.1",
    "ipywidgets==8.1.7",
]

[dependency-groups]
dev = ["mypy>=1.14.1", "pre-commit>=4.0.1", "ruff>=0.9.0"]

[tool.mypy]
disable_error_code = ["import-untyped"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
